+++ erica-domain/src/main/java/com/matttax/erica/domain/config/MoveWordsRequest.kt
@@ -1,7 +1,7 @@
 package com.matttax.erica.domain.config
 
 data class MoveWordsRequest(
-    val idFrom: Int,
+    val idFrom: Long,
     val idTo: Long,
     val words: List<Long>
 )
