+++ erica-domain/src/main/java/com/matttax/erica/domain/config/SetId.kt
@@ -2,6 +2,6 @@ package com.matttax.erica.domain.config
 
 sealed class SetId {
     object None: SetId()
-    data class One(val id: Int): SetId()
-    class Many(vararg val ids: Int): SetId()
+    data class One(val id: Long): SetId()
+    class Many(vararg val ids: Long): SetId()
 }
